-Selected image thing added. -I managed to get a prototype Shuffle button to work. It'd be awesome if Chris confirmed and cleaned and commented everything though.
Seriously though, game.py is already bigger than 1000 lines, it desperately needs a second opinion, and you need to learn how this works anyway. Please decide to make an attempt on cleaning and commenting, let me know (So I don't continue making epic progress, even while on vacation), and do it. You don't even need to touch the text parts. It's critical that we do that before it gets out of hand, and that's not very far off.

Next:
-Assuming Chris does what he can regarding the last commit's notes and finishes his cleanup (And I know he will), the next step would be the custom bar.
--Bar needs to be ripped.
--Please direct me to a draw function in Pyglet that will allow me to draw x amount of bars in it with a specific color.

After:
-Remove all text functionality.
--Hide the shell. How would we do that?
-Animations.
-Assess lag. This should be able to run faster, and whether it's the concept or the implementation that is slowing it down, something must be done.
-Continuing working with chips and handling other online capabilities.